Extract HiveStatisticsProvider
