+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveMetadata.java
@@ -14,7 +14,6 @@
 package com.facebook.presto.hive;
 
 import com.facebook.presto.hive.metastore.Column;
-import com.facebook.presto.hive.metastore.ColumnStatistics;
 import com.facebook.presto.hive.metastore.Database;
 import com.facebook.presto.hive.metastore.HivePrivilegeInfo;
 import com.facebook.presto.hive.metastore.HivePrivilegeInfo.HivePrivilege;
@@ -24,6 +23,7 @@ import com.facebook.presto.hive.metastore.SemiTransactionalHiveMetastore;
 import com.facebook.presto.hive.metastore.SemiTransactionalHiveMetastore.WriteMode;
 import com.facebook.presto.hive.metastore.StorageFormat;
 import com.facebook.presto.hive.metastore.Table;
+import com.facebook.presto.hive.statistics.HiveStatisticsProvider;
 import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ColumnMetadata;
 import com.facebook.presto.spi.ConnectorInsertTableHandle;
@@ -50,12 +50,10 @@ import com.facebook.presto.spi.predicate.Domain;
 import com.facebook.presto.spi.predicate.NullableValue;
 import com.facebook.presto.spi.predicate.TupleDomain;
 import com.facebook.presto.spi.security.Privilege;
-import com.facebook.presto.spi.statistics.Estimate;
 import com.facebook.presto.spi.statistics.TableStatistics;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeManager;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Preconditions;
 import com.google.common.base.Verify;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -72,23 +70,18 @@ import org.apache.hadoop.hive.ql.exec.FileSinkOperator;
 import org.apache.hadoop.mapred.JobConf;
 import org.joda.time.DateTimeZone;
 
-import javax.annotation.Nullable;
-
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.OptionalDouble;
 import java.util.OptionalLong;
 import java.util.Properties;
 import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
-import java.util.stream.DoubleStream;
 
 import static com.facebook.presto.hive.HiveColumnHandle.BUCKET_COLUMN_NAME;
 import static com.facebook.presto.hive.HiveColumnHandle.ColumnType.HIDDEN;
@@ -176,6 +169,7 @@ public class HiveMetadata
     private final HiveStorageFormat defaultStorageFormat;
     private final TypeTranslator typeTranslator;
     private final String prestoVersion;
+    private final HiveStatisticsProvider hiveStatisticsProvider;
 
     public HiveMetadata(
             String connectorId,
@@ -192,7 +186,8 @@ public class HiveMetadata
             TableParameterCodec tableParameterCodec,
             JsonCodec<PartitionUpdate> partitionUpdateCodec,
             TypeTranslator typeTranslator,
-            String prestoVersion)
+            String prestoVersion,
+            HiveStatisticsProvider hiveStatisticsProvider)
     {
         this.connectorId = requireNonNull(connectorId, "connectorId is null");
 
@@ -211,6 +206,7 @@ public class HiveMetadata
         this.defaultStorageFormat = requireNonNull(defaultStorageFormat, "defaultStorageFormat is null");
         this.typeTranslator = requireNonNull(typeTranslator, "typeTranslator is null");
         this.prestoVersion = requireNonNull(prestoVersion, "prestoVersion is null");
+        this.hiveStatisticsProvider = requireNonNull(hiveStatisticsProvider, "hiveStatisticsProvider is null");
     }
 
     public SemiTransactionalHiveMetastore getMetastore()
@@ -366,145 +362,7 @@ public class HiveMetadata
 
         Map<String, ColumnHandle> tableColumns = getTableColumnNamesToHandlesMap(hivePartitions, session);
 
-        Map<String, PartitionStatistics> partitionStatisticsMap = getPartitionsStatistics(hivePartitions, tableColumns.keySet());
-
-        TableStatistics.Builder tableStatistics = TableStatistics.builder();
-
-        tableStatistics.setRowCount(calculateRowsCount(partitionStatisticsMap));
-
-        for (Map.Entry<String, ColumnHandle> columnEntry : tableColumns.entrySet()) {
-            String columnName = columnEntry.getKey();
-            HiveColumnHandle hiveColumnHandle = (HiveColumnHandle) columnEntry.getValue();
-            if (getColumnMetadata(session, tableHandle, hiveColumnHandle).isHidden()) {
-                continue;
-            }
-            com.facebook.presto.spi.statistics.ColumnStatistics.Builder columnStatistics = com.facebook.presto.spi.statistics.ColumnStatistics.builder();
-            if (hiveColumnHandle.isPartitionKey()) {
-                columnStatistics.setDistinctValuesCount(calculateDistinctValuesCountForPartitioningKey(hiveColumnHandle, hivePartitions));
-                columnStatistics.setNullsCount(calculateNullsCountForPartitioningKey(hiveColumnHandle, hivePartitions, partitionStatisticsMap));
-            }
-            else {
-                columnStatistics.setDistinctValuesCount(calculateDistinctValuesCount(partitionStatisticsMap, columnName));
-                columnStatistics.setNullsCount(calculateNullsCount(partitionStatisticsMap, columnName));
-            }
-            tableStatistics.setColumnStatistics(hiveColumnHandle, columnStatistics.build());
-        }
-        return tableStatistics.build();
-    }
-
-    private Estimate calculateRowsCount(Map<String, PartitionStatistics> partitionStatisticsMap)
-    {
-        List<Long> knownPartitionRowCounts = partitionStatisticsMap.values().stream()
-                .map(PartitionStatistics::getRowCount)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .collect(toList());
-
-        long knownPartitionRowCountsSum = knownPartitionRowCounts.stream().mapToLong(a -> a).sum();
-        long partitionsWithStatsCount = knownPartitionRowCounts.size();
-        long allPartitionsCount = partitionStatisticsMap.size();
-
-        if (partitionsWithStatsCount == 0) {
-            return Estimate.unknownValue();
-        }
-        return new Estimate(1.0 * knownPartitionRowCountsSum / partitionsWithStatsCount * allPartitionsCount);
-    }
-
-    private Estimate calculateDistinctValuesCount(Map<String, PartitionStatistics> statisticsByPartitionName, String column)
-    {
-        return getColumnStatisticsValue(
-                statisticsByPartitionName,
-                column,
-                columnStatistics -> {
-                    if (columnStatistics.getDistinctValuesCount().isPresent()) {
-                        return OptionalDouble.of(columnStatistics.getDistinctValuesCount().getAsLong());
-                    }
-                    else {
-                        return OptionalDouble.empty();
-                    }
-                },
-                DoubleStream::max);
-    }
-
-    private Estimate calculateNullsCount(Map<String, PartitionStatistics> statisticsByPartitionName, String column)
-    {
-        return getColumnStatisticsValue(
-                statisticsByPartitionName,
-                column,
-                columnStatistics -> {
-                    if (columnStatistics.getNullsCount().isPresent()) {
-                        return OptionalDouble.of(columnStatistics.getNullsCount().getAsLong());
-                    }
-                    else {
-                        return OptionalDouble.empty();
-                    }
-                },
-                doubleStream -> {
-                    List<Double> values = doubleStream.mapToObj(Double::valueOf).collect(toList());
-                    long partitionsWithStatsCount = values.size();
-                    double valuesSum = values.stream().mapToDouble(Double::doubleValue).sum();
-                    if (partitionsWithStatsCount == 0) {
-                        return OptionalDouble.empty();
-                    }
-                    else {
-                        int allPartitionsCount = statisticsByPartitionName.size();
-                        return OptionalDouble.of(allPartitionsCount / partitionsWithStatsCount * valuesSum);
-                    }
-                });
-    }
-
-    private Estimate calculateDistinctValuesCountForPartitioningKey(HiveColumnHandle hiveColumnHandle, List<HivePartition> partitions)
-    {
-        return new Estimate(partitions.stream()
-                .map(partition -> partition.getKeys().get(hiveColumnHandle))
-                .collect(toSet())
-                .size());
-    }
-
-    private Estimate calculateNullsCountForPartitioningKey(HiveColumnHandle hiveColumnHandle, List<HivePartition> hivePartitions, Map<String, PartitionStatistics> partitionStatisticsMap)
-    {
-        List<Long> knownPartitionRowCounts = partitionStatisticsMap.values().stream()
-                .map(PartitionStatistics::getRowCount)
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .collect(toList());
-
-        if (knownPartitionRowCounts.isEmpty()) {
-            return Estimate.unknownValue();
-        }
-
-        long knownPartitionRowCountsSum = knownPartitionRowCounts.stream().mapToLong(Long::longValue).sum();
-        long averagePartitionRowsCount = knownPartitionRowCountsSum / knownPartitionRowCounts.size();
-
-        return new Estimate(hivePartitions.stream()
-                .filter(partition -> partition.getKeys().get(hiveColumnHandle).isNull())
-                .map(HivePartition::getPartitionId)
-                .mapToLong(partitionId -> partitionStatisticsMap.get(partitionId).getRowCount().orElse(averagePartitionRowsCount))
-                .sum());
-    }
-
-    private Estimate getColumnStatisticsValue(
-            Map<String, PartitionStatistics> statisticsByPartitionName,
-            String column,
-            Function<ColumnStatistics, OptionalDouble> valueExtractFunction,
-            Function<DoubleStream, OptionalDouble> valueAggregateFunction)
-    {
-        DoubleStream intermediateStream = statisticsByPartitionName.values().stream()
-                .map(PartitionStatistics::getColumnStatistics)
-                .filter(columnStatisticsMap -> columnStatisticsMap.containsKey(column))
-                .map(columnStatisticsMap -> columnStatisticsMap.get(column))
-                .map(valueExtractFunction)
-                .filter(OptionalDouble::isPresent)
-                .mapToDouble(OptionalDouble::getAsDouble);
-
-        OptionalDouble statisicsValue = valueAggregateFunction.apply(intermediateStream);
-
-        if (statisicsValue.isPresent()) {
-            return new Estimate(statisicsValue.getAsDouble());
-        }
-        else {
-            return Estimate.unknownValue();
-        }
+        return hiveStatisticsProvider.getTableStatistics(session, tableHandle, hivePartitions, tableColumns);
     }
 
     private Map<String, ColumnHandle> getTableColumnNamesToHandlesMap(List<HivePartition> partitions, ConnectorSession session)
@@ -515,95 +373,6 @@ public class HiveMetadata
         }
 
         return ImmutableMap.copyOf(getColumnHandles(session, getTableHandle(session, hivePartition.getTableName())));
-    }
-
-    private Map<String, PartitionStatistics> getPartitionsStatistics(List<HivePartition> hivePartitions, Set<String> tableColumns)
-    {
-        if (hivePartitions.isEmpty()) {
-            return ImmutableMap.of();
-        }
-        boolean unpartitioned = hivePartitions.stream().anyMatch(partition -> partition.getPartitionId().equals(HivePartition.UNPARTITIONED_ID));
-        if (unpartitioned) {
-            checkArgument(hivePartitions.size() == 1, "expected only one hive partition");
-        }
-        SchemaTableName tableName = getTableName(hivePartitions);
-
-        if (unpartitioned) {
-            return ImmutableMap.of(HivePartition.UNPARTITIONED_ID, getUnpartitionedStatistics(tableName, tableColumns));
-        }
-        else {
-            return getPartitionedStatistics(tableName, hivePartitions, tableColumns);
-        }
-    }
-
-    private SchemaTableName getTableName(List<HivePartition> hivePartitions)
-    {
-        Set<SchemaTableName> tableNames = hivePartitions.stream().map(HivePartition::getTableName).collect(toSet());
-        Preconditions.checkArgument(tableNames.size() == 1, "all hive partitions must have same table name; got %s", tableNames);
-        return Iterables.getOnlyElement(tableNames);
-    }
-
-    private PartitionStatistics getUnpartitionedStatistics(SchemaTableName schemaTableName, Set<String> tableColumns)
-    {
-        String databaseName = schemaTableName.getSchemaName();
-        String tableName = schemaTableName.getTableName();
-        Table table = metastore.getTable(databaseName, tableName)
-                .orElseThrow(() -> new IllegalArgumentException(format("Could not get metadata for table %s.%s", databaseName, tableName)));
-
-        Map<String, ColumnStatistics> tableColumnStatistics = metastore.getTableColumnStatistics(databaseName, tableName, tableColumns).orElse(ImmutableMap.of());
-
-        return readStatisticsFromParameters(table.getParameters(), tableColumnStatistics);
-    }
-
-    private Map<String, PartitionStatistics> getPartitionedStatistics(SchemaTableName schemaTableName, List<HivePartition> hivePartitions, Set<String> tableColumns)
-    {
-        String databaseName = schemaTableName.getSchemaName();
-        String tableName = schemaTableName.getTableName();
-
-        ImmutableMap.Builder<String, PartitionStatistics> resultMap = ImmutableMap.builder();
-
-        List<String> partitionNames = hivePartitions.stream().map(HivePartition::getPartitionId).collect(Collectors.toList());
-        Map<String, Map<String, ColumnStatistics>> partitionColumnStatisticsMap =
-                metastore.getPartitionColumnStatistics(databaseName, tableName, new HashSet<>(partitionNames), tableColumns)
-                        .orElse(ImmutableMap.of());
-
-        Map<String, Optional<Partition>> partitionsByNames = metastore.getPartitionsByNames(databaseName, tableName, partitionNames);
-        for (String partitionName : partitionNames) {
-            Map<String, String> partitionParameters = partitionsByNames.get(partitionName)
-                    .map(Partition::getParameters)
-                    .orElseThrow(() -> new IllegalArgumentException(format("Could not get metadata for partition %s.%s.%s", databaseName, tableName, partitionName)));
-            Map<String, ColumnStatistics> partitionColumnStatistics = partitionColumnStatisticsMap.getOrDefault(partitionName, ImmutableMap.of());
-            resultMap.put(partitionName, readStatisticsFromParameters(partitionParameters, partitionColumnStatistics));
-        }
-
-        return resultMap.build();
-    }
-
-    private PartitionStatistics readStatisticsFromParameters(Map<String, String> parameters, Map<String, ColumnStatistics> columnStatistics)
-    {
-        boolean columnStatsAcurate = Boolean.valueOf(Optional.ofNullable(parameters.get("COLUMN_STATS_ACCURATE")).orElse("false"));
-        Optional<Long> numFiles = convertStringParameter(parameters.get("numFiles"));
-        Optional<Long> numRows = convertStringParameter(parameters.get("numRows"));
-        Optional<Long> rawDataSize = convertStringParameter(parameters.get("rawDataSize"));
-        Optional<Long> totalSize = convertStringParameter(parameters.get("totalSize"));
-        return new PartitionStatistics(columnStatsAcurate, numFiles, numRows, rawDataSize, totalSize, columnStatistics);
-    }
-
-    private Optional<Long> convertStringParameter(@Nullable String parameterValue)
-    {
-        if (parameterValue == null) {
-            return Optional.empty();
-        }
-        try {
-            long longValue = Long.parseLong(parameterValue);
-            if (longValue < 0) {
-                return Optional.empty();
-            }
-            return Optional.of(longValue);
-        }
-        catch (NumberFormatException e) {
-            return Optional.empty();
-        }
     }
 
     private List<SchemaTableName> listTables(ConnectorSession session, SchemaTablePrefix prefix)
